refactor: 패키지명 소문자 컨벤션 적용하여 수정
