+++ src/test/java/mission/lotto/domain/LottoNumbersTest.java
@@ -1,4 +1,4 @@
-package LottoMission.domain;
+package mission.lotto.domain;
 
 import org.junit.jupiter.api.Test;
