+++ src/test/java/mission/lotto/util/RandomUtilTest.java
@@ -1,4 +1,4 @@
-package LottoMission.util;
+package mission.lotto.util;
 
 import org.junit.jupiter.api.Test;
